+++ session_cache.go
@@ -1,6 +1,10 @@
 package pirsch
 
-import "time"
+import (
+	"fmt"
+	"sync"
+	"time"
+)
 
 // SessionCache caches sessions.
 type SessionCache interface {
@@ -12,4 +16,11 @@ type SessionCache interface {
 
 	// Clear clears the cache.
 	Clear()
+
+	// NewMutex creates a new mutex for given client ID and fingerprint.
+	NewMutex(uint64, uint64) sync.Locker
+}
+
+func getSessionKey(clientID, fingerprint uint64) string {
+	return fmt.Sprintf("%d_%d", clientID, fingerprint)
 }